+++ apps/menu-service/src/main/java/logiroute/logiroute_menu/dto/MenuItemResponse.java
@@ -8,7 +8,7 @@ public record MenuItemResponse(
         UUID categoryId,
         String name,
         String description,
-        Long basePriceCents,
+        Integer basePriceCents,
         Boolean isVegan,
         Boolean isGlutenFree,
         Boolean isAvailable,
